prac-5/requirements.txt
pysimplegui==4.60.4
Pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.3